[ucsim/stm8] Fix timer IRQ generation
The redesign of the IRQ handling adds hw watchers to the cells
containing interrupt flags. This means that interrupt flags MUST be set
via write* methods rather than set* methods as only the write* methods
notify the hw watchers of the the update. More generally, there can be
arbitrary numbers of hw watchers on a cell (think of vcd for example) so
ALL updates should be via write* methods unless you have a very good
reason to hide them. Possibly the set* methods should be made private
and/or removed?

Signed-off-by: Mike Jagdis <mjagdis@eris-associates.co.uk> (github: mjagdis)