[ucsim] Add smart dump mode
Unless a dump mode is explicitly specified dump operates in "smart"
mode. In this mode each address is looked up in the var table. If a var
exists for it and has high/low bits defined the address is taken to be a
register and dumped on a line of its own in binary, hex, character and
decimal formats. Otherwise if a var exists that has no bits defined the
var name is shown as a label at that address and data is dumped in hex.

The width used for the label field is controlled by the "label_width"
option. If set to -1 (the default) the length of the longest variable
name is used instead.

Signed-off-by: Mike Jagdis <mjagdis@eris-associates.co.uk> (github: mjagdis)