[ucsim] Use line limit in smart dump mode output
It makes no sense to default the end address based on bytes-per-line and
the intended line count when outputting in smart dump mode. Instead we
keep track of how many lines have been output and stop when we have
output enough. There may, however, be more lines output than intended
because we always finish the address we are on and it may have multiple
labels or multiple named bits.

Signed-off-by: Mike Jagdis <mjagdis@eris-associates.co.uk> (github: mjagdis)