[ucsim] Support for "stretch" instructions.
Not all instructions have well defined execution time, for example
wait-for-interrupt or wait-for-event instructions. Have the simulator
handle these by allowing execution of an instruction to return
resNOT_DONE as an indication that the PC should not be updated and that
the next step while resume the instruction. This allows long running
instructions to tick hardware and be interruptible from the console.

Signed-off-by: Mike Jagdis <mjagdis@eris-associates.co.uk>